minor code cleanup
git-svn-id: file:///var/lib/svn/clamav-devel/trunk/clamav-devel@2224 77e5149b-7576-45b1-b177-96237e5ba77b

+++ shared/output.c
@@ -93,7 +93,7 @@ int mdprintf(int desc, const char *str, ...)
     if(bytes == -1)
 	return bytes;
 
-    if(bytes >= sizeof(buff))
+    if(bytes >= (int) sizeof(buff))
 	bytes = sizeof(buff) - 1;
 
     return send(desc, buff, bytes, 0);
